Moved files to analyze.computation
