+++ src/main/java/model/Group.java
@@ -159,7 +159,7 @@ public class Group extends HashMap<String, RecordList> {
 
 		for (String filePath : this.keySet()) {
 			for (Record record : this.get(filePath)) {
-				RecordField id = (RecordField) record.get(primary);
+				DataField id = (DataField) record.get(primary);
 				if (res.containsKey(id.toString())) {
 					RecordList list = res.get(id.toString());
 					list.add(record);
