+++ src/test/java/analyze/computing/ComputerTest.java
@@ -1,4 +1,4 @@
-package analyze;
+package analyze.computing;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
