+++ src/test/java/model/RecordFieldDoubleTest.java
@@ -8,7 +8,7 @@ public class RecordFieldDoubleTest {
 
 	@Test
 	public void toStringTest() {
-		RecordFieldDouble fl = new RecordFieldDouble(10.0);
+	    DataFieldDouble fl = new DataFieldDouble(10.0);
 		assertEquals(fl.toString(), "10.0");
 	}
