+++ src/test/java/model/RecordFieldIntTest.java
@@ -8,7 +8,7 @@ public class RecordFieldIntTest {
 
 	@Test
 	public void toStringTest() {
-		RecordFieldInt fl = new RecordFieldInt(10);
+	    DataFieldInt fl = new DataFieldInt(10);
 		assertEquals(fl.toString(), "10");
 	}
